+++ tests/image/embedding/test_model.py
@@ -87,3 +87,59 @@ def test_vissl_training(backbone, training_strategy, head, pretraining_transform
     for prediction_batch in predictions:
         for prediction in prediction_batch:
             assert prediction.size(0) == embedding_size
+
+
+@pytest.mark.skipif(not (_IMAGE_AVAILABLE and _VISSL_AVAILABLE), reason="vissl not installed.")
+@pytest.mark.parametrize(
+    "backbone, training_strategy, head, pretraining_transform, expected_exception",
+    [
+        ("resnet18", "simclr", "simclr_head", None, ValueError),
+        ("resnet18", "simclr", None, "simclr_transform", KeyError),
+    ],
+)
+def test_vissl_training_with_wrong_arguments(
+    backbone, training_strategy, head, pretraining_transform, expected_exception
+):
+    with pytest.raises(expected_exception):
+        ImageEmbedder(
+            backbone=backbone,
+            training_strategy=training_strategy,
+            head=head,
+            pretraining_transform=pretraining_transform,
+        )
+
+
+@pytest.mark.skipif(not _IMAGE_AVAILABLE, reason="torch vision not installed.")
+@pytest.mark.parametrize(
+    "backbone, embedding_size",
+    [
+        ("resnet18", 512),
+        ("vit_small_patch16_224", 384),
+    ],
+)
+def test_only_embedding(backbone, embedding_size):
+    datamodule = ImageClassificationData.from_datasets(
+        predict_dataset=FakeData(8),
+        batch_size=4,
+        transform_kwargs=dict(image_size=(224, 224)),
+    )
+
+    embedder = ImageEmbedder(backbone=backbone)
+    trainer = flash.Trainer()
+
+    predictions = trainer.predict(embedder, datamodule=datamodule)
+    for prediction_batch in predictions:
+        for prediction in prediction_batch:
+            assert prediction.size(0) == embedding_size
+
+
+@pytest.mark.skipif(not _IMAGE_AVAILABLE, reason="torch vision not installed.")
+def test_not_implemented_steps():
+    embedder = ImageEmbedder(backbone="resnet18")
+
+    with pytest.raises(NotImplementedError):
+        embedder.training_step([], 0)
+    with pytest.raises(NotImplementedError):
+        embedder.validation_step([], 0)
+    with pytest.raises(NotImplementedError):
+        embedder.test_step([], 0)